requirements.txt
pandas>=2.0.0
numpy>=1.24.0
akshare>=1.12.0
akshare-proxy-patch>=0.2.13
baostock>=0.8.9
requests>=2.28.0
scipy>=1.10.0
scikit-learn>=1.3.0
plotly>=5.18.0
deepagents>=0.4.0
langchain-openai>=0.3.0
pyyaml>=6.0
jqdatasdk>=1.9.0
